% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Qm3s.R
\name{as.Qm3s}
\alias{as.Qm3s}
\title{Coerce \link{data.frame} or content of a \link{data.frame} into a \emph{Qm3s} object ready
for plotting}
\usage{
as.Qm3s(...)
}
\arguments{
\item{...}{A \link{data.frame} for a single argument, or the arguments of function \link{data.frame}}
}
\value{
A \link{data.frame} of class \emph{Qm3s}
}
\description{
Coerce \link{data.frame} or content of a \link{data.frame} into a \emph{Qm3s} object ready
for plotting
}
