% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGRiwrm.R
\name{CheckColumnTypes}
\alias{CheckColumnTypes}
\title{Check of the column types of a \link{data.frame}}
\usage{
CheckColumnTypes(df, coltypes)
}
\arguments{
\item{df}{\link{data.frame} to check}

\item{coltypes}{named \link{list} with the name of the columns to check as key and the required type as value}
}
\value{
\link{NULL} or error message if a wrong type is detected
}
\description{
Check of the column types of a \link{data.frame}
}
\examples{
CheckColumnTypes(
  data.frame(string = c("A"), numeric = c(1), stringsAsFactors = FALSE),
  list(string = "character", numeric = "double")
)

}
