% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkRunModelParameters}
\alias{checkRunModelParameters}
\title{Check the parameters of RunModel methods}
\usage{
checkRunModelParameters(InputsModel, RunOptions, Param)
}
\arguments{
\item{InputsModel}{a \code{GRiwrmInputsModel} object (See \link{CreateInputsModel.GRiwrm})}

\item{RunOptions}{a \code{GRiwrmRunOptions} object (See \link{CreateRunOptions.GRiwrmInputsModel})}

\item{Param}{a \link{list} of \link{numeric} containing model parameters of each node of the network}
}
\description{
Stop the execution if an error is detected.
}
