% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCalibOptions.InputsModel.R
\name{CreateCalibOptions.InputsModel}
\alias{CreateCalibOptions.InputsModel}
\title{Wrapper to \link[airGR:CreateCalibOptions]{airGR::CreateCalibOptions} for one sub-basin.}
\usage{
\method{CreateCalibOptions}{InputsModel}(InputsModel, ...)
}
\arguments{
\item{InputsModel}{object of class \emph{InputsModel}. See \link[airGR:CreateInputsModel]{airGR::CreateInputsModel} for details}

\item{...}{Arguments passed to \link[airGR:CreateCalibOptions]{airGR::CreateCalibOptions}}
}
\description{
Wrapper to \link[airGR:CreateCalibOptions]{airGR::CreateCalibOptions} for one sub-basin.
}
