% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInputsModel.GRiwrm.R
\name{CreateEmptyGRiwrmInputsModel}
\alias{CreateEmptyGRiwrmInputsModel}
\title{Create an empty InputsModel object for \strong{airGRiwrm} nodes}
\usage{
CreateEmptyGRiwrmInputsModel(griwrm)
}
\arguments{
\item{griwrm}{a \code{GRiwrm} object (See \link{CreateGRiwrm})}
}
\value{
\emph{GRiwrmInputsModel} empty object
}
\description{
Create an empty InputsModel object for \strong{airGRiwrm} nodes
}
