% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInputsModel.default.R
\name{CreateInputsModel.default}
\alias{CreateInputsModel.default}
\title{Wrapper for \link[airGR:CreateInputsModel]{airGR::CreateInputsModel} for one sub-basin.}
\usage{
\method{CreateInputsModel}{default}(x, ...)
}
\arguments{
\item{x}{\link{function} hydrological model function (e.g. \link[airGR:RunModel_GR4J]{airGR::RunModel_GR4J}...)}

\item{...}{arguments passed to \link[airGR:CreateInputsModel]{airGR::CreateInputsModel}}
}
\description{
Wrapper for \link[airGR:CreateInputsModel]{airGR::CreateInputsModel} for one sub-basin.
}
