% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInputsModel.R
\name{CreateInputsModel}
\alias{CreateInputsModel}
\title{Generic function for creating \code{InputsModel} object for either \strong{airGR} or \strong{airGRiwrm}}
\usage{
CreateInputsModel(x, ...)
}
\arguments{
\item{x}{First parameter determining which InputsModel object is created}

\item{...}{further arguments passed to or from other methods.}
}
\value{
InputsModel or GRiwrmInputsObject object
}
\description{
See the methods \link{CreateInputsModel.GRiwrm} for \strong{airGRiwrm} and \link{CreateInputsModel.default} for \strong{airGR}.
}
