% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInputsModel.GRiwrm.R
\name{CreateOneGRiwrmInputsModel}
\alias{CreateOneGRiwrmInputsModel}
\title{Create one InputsModel for a \strong{airGRiwrm} node}
\usage{
CreateOneGRiwrmInputsModel(id, griwrm, DatesR, Precip, PotEvap, Qobs)
}
\arguments{
\item{id}{string of the node identifier}

\item{griwrm}{See \link{CreateGRiwrm})}

\item{DatesR}{vector of dates required to create the GR model and CemaNeige module inputs}

\item{Precip}{time series of potential evapotranspiration (catchment average) (mm/time step)}

\item{PotEvap}{time series of potential evapotranspiration (catchment average) (mm/time step)}

\item{Qobs}{Matrix or data frame of numeric containing observed flow (mm/time step). Column names correspond to node IDs}
}
\value{
\emph{InputsModel} object for one.
}
\description{
Create one InputsModel for a \strong{airGRiwrm} node
}
