% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRunOptions.GRiwrmInputsModel.R
\name{CreateRunOptions.GRiwrmInputsModel}
\alias{CreateRunOptions.GRiwrmInputsModel}
\title{Create \emph{GRiwrmRunOptions} object for running and calibrating model in \strong{airGRiwrm}.}
\usage{
\method{CreateRunOptions}{GRiwrmInputsModel}(InputsModel, ...)
}
\arguments{
\item{InputsModel}{object of class \emph{GRiwrmInputsModel}, see \code{\link{CreateInputsModel.GRiwrm}} for details.}

\item{...}{further arguments passed to \code{\link[airGR]{CreateRunOptions}}.}
}
\value{
\emph{GRiwrmRunOptions} object for running and calibrating model in \strong{airGRiwrm}.
}
\description{
Create \emph{GRiwrmRunOptions} object for running and calibrating model in \strong{airGRiwrm}.
}
