% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRunOptions.InputsModel.R
\name{CreateRunOptions.InputsModel}
\alias{CreateRunOptions.InputsModel}
\title{Wrapper for \link[airGR:CreateRunOptions]{airGR::CreateRunOptions} for one sub-basin.}
\usage{
\method{CreateRunOptions}{InputsModel}(InputsModel, ...)
}
\arguments{
\item{InputsModel}{object of class \emph{InputsModel}, see \link[airGR:CreateInputsModel]{airGR::CreateInputsModel} for details.}

\item{...}{Arguments passed to \link[airGR:CreateRunOptions]{airGR::CreateRunOptions}}
}
\description{
Wrapper for \link[airGR:CreateRunOptions]{airGR::CreateRunOptions} for one sub-basin.
}
