% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagramGRiwrm.R
\name{DiagramGRiwrm}
\alias{DiagramGRiwrm}
\title{Display a diagram representing the network structure of a GRiwrm object}
\usage{
DiagramGRiwrm(griwrm, display = TRUE, orientation = "LR")
}
\arguments{
\item{griwrm}{the GRiwrm object to display.}

\item{display}{if \code{TRUE} displays the diagram with \code{DiagrammeR::mermaid}, return the mermaid code otherwise.}

\item{orientation}{Orientation of the graph. "LR" by default.}
}
\value{
Mermaid code of the diagram id display is \code{FALSE}.
}
\description{
Display a diagram representing the network structure of a GRiwrm object
}
\details{
This function only works inside RStudio because the HTMLwidget produced by DiagrammeR
is not handled on some platforms
}
\examples{
\dontrun{
# Display diagram
DiagramGRiwrm(griwrm)
# Is the same as
DiagrammeR::mermaid(DiagramGRiwrm(griwrm, display = FALSE), width = "100\%", height = "100\%")
}

}
