% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doSupervision}
\alias{doSupervision}
\title{Supervision for the current time step}
\usage{
doSupervision(supervisor)
}
\arguments{
\item{supervisor}{\code{Supervisor} (See \link{CreateSupervisor})}
}
\description{
Supervision for the current time step
}
