% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.Calibration.R
\name{extractParam}
\alias{extractParam}
\alias{extractParam.GRiwrmOutputsCalib}
\alias{extractParam.GRiwrmOutputsModel}
\title{Extract calibrated parameters}
\usage{
extractParam(x)

\method{extractParam}{GRiwrmOutputsCalib}(x)

\method{extractParam}{GRiwrmOutputsModel}(x)
}
\arguments{
\item{x}{[\Object of class \emph{GRiwrmOutputsModel}] returned by [Calibration.GRiwrmInputsModel]

[\Object of class \emph{GRiwrmOutputsModel}]: R:\%5C\%5CObject\%20of\%20class\%20\emph{GRiwrmOutputsModel}\%5C
[Calibration.GRiwrmInputsModel]: R:Calibration.GRiwrmInputsModel}
}
\value{
[\list] [numeric] [vector] containing the calibrated parameters
of each modeled node.

[\list]: R:\%5C\%5Clist\%5C
[numeric]: R:numeric
[vector]: R:vector
}
\description{
Extract \link{list} of parameters from the output of \link{Calibration.GRiwrmInputsModel}
which can be directly used as argument \code{Param} of \link{RunModel.GRiwrmInputsModel}
and \link{RunModel.Supervisor}.
}
\details{
See vignettes and example of \link{RunModel_Reservoir} for examples of use.
}
\seealso{
\link{Calibration}, \link{RunModel.GRiwrmInputsModel}, \link{RunModel.Supervisor}
}
