% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDataFromLocation}
\alias{getDataFromLocation}
\title{Retrieve data in the model for the current time steps}
\usage{
getDataFromLocation(loc, sv)
}
\arguments{
\item{loc}{location of the data}

\item{sv}{a \code{Supervisor} (See \link{CreateSupervisor})}
}
\value{
\link{numeric} retrieved data at the location
}
\description{
This function should be call inside a Supervisor
}
