% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInputsCrit.GRiwrmInputsModel.R
\name{getDefaultAprioriIds}
\alias{getDefaultAprioriIds}
\title{Get default AprioriIds from direct upstream nodes of each node}
\usage{
getDefaultAprioriIds(InputsModel)
}
\arguments{
\item{InputsModel}{object of class \emph{InputsModel} or \emph{GRiwrmInputsModel}. See \link{CreateInputsModel}}
}
\value{
A \link{list} named with node Ids and containing the Ids of the upstream
nodes that can be used for apriori parameters.
}
\description{
Get default AprioriIds from direct upstream nodes of each node
}
