% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInputsModel.GRiwrm.R
\name{getModelTimeStep}
\alias{getModelTimeStep}
\title{Check time steps of the model of all the nodes and return the time step in seconds}
\usage{
getModelTimeStep(InputsModel)
}
\arguments{
\item{InputsModel}{a \code{GRiwrmInputsModel}}
}
\value{
A \link{numeric} representing the time step in seconds
}
\description{
This function is called inside \link{CreateInputsModel.GRiwrm} for defining the time step of the big model.
}
