% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGRiwrm.R
\name{getNodeRanking}
\alias{getNodeRanking}
\title{Sorting of the nodes from upstream to downstream}
\usage{
getNodeRanking(griwrm)
}
\arguments{
\item{griwrm}{[object of class \code{GRiwrm}] see \link{CreateGRiwrm} for details}
}
\value{
\link{numeric} ordered node names
}
\description{
Sorting of the nodes from upstream to downstream
}
