% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getNoSD_Ids}
\alias{getNoSD_Ids}
\title{Id of sub-basins not using SD model}
\usage{
getNoSD_Ids(InputsModel)
}
\arguments{
\item{InputsModel}{\code{GRiwrmInputsModel} object}
}
\value{
\link{character} IDs of the sub-basins not using SD model
}
\description{
Id of sub-basins not using SD model
}
