% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.GRiwrm.R
\name{getSD_Ids}
\alias{getSD_Ids}
\alias{getNoSD_Ids}
\title{Function to get the IDs of sub-basins using SD model or not}
\usage{
getSD_Ids(InputsModel, add_diversions = FALSE)

getNoSD_Ids(InputsModel, include_diversion = TRUE)
}
\arguments{
\item{InputsModel}{\link[=CreateInputsModel.GRiwrm]{GRiwrmInputsModel} object}

\item{add_diversions}{\link{logical} for adding upstream nodes due to diversion}

\item{include_diversion}{\link{logical} for including diversion nodes}
}
\value{
\link{character} IDs of the sub-basins using SD model or not
}
\description{
Function to get the IDs of sub-basins using SD model or not
}
