% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRiwrm.R
\name{GRiwrm}
\alias{GRiwrm}
\title{Generate a network description containing all hydraulic nodes and the description
of their connections}
\usage{
GRiwrm(
  db,
  cols = list(id = "id", down = "down", length = "length", model = "model", area =
    "area"),
  keep_all = FALSE
)
}
\arguments{
\item{db}{a \link{data.frame} containing the description of the network (See details)}

\item{cols}{named list or vector for matching columns of \code{db} parameter. By default, mandatory columns names are: \code{id}, \code{down}, \code{length}. But other names can be handled with a named list or vector containing items defined as \code{"required name" = "column name in db"}}

\item{keep_all}{keep all column of \code{db} or keep only columns defined in \code{cols}}
}
\value{
\code{GRiwrm} class object containing the description of diagram of the semi-distributed catchment model
}
\description{
Generate a network description containing all hydraulic nodes and the description
of their connections
}
\details{
\code{db} is a \link{data.frame} which at least contains in its columns:
\itemize{
\item a node identifier (column \code{id}),
\item the identifier and the hydraulic distance to the downstream node (\link{character} columns \code{down} and \link{numeric} columns \code{length} in meters). The last downstream node should have fields \code{down} and \code{length} set to \code{NA},
\item the area of the basin (\link{numeric} column \code{area} in km^2^)
\item the hydrological model to use if so (\link{character} column \code{model}).
}
}
