% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.RunModel.R
\name{merge.OutputsModel}
\alias{merge.OutputsModel}
\alias{merge.GRiwrmOutputsModel}
\title{Merge Two outputs of airGR simulations}
\usage{
\method{merge}{OutputsModel}(x, y, ...)

\method{merge}{GRiwrmOutputsModel}(x, y, ...)
}
\arguments{
\item{x, y}{\strong{OutputsModel} objects from \link[airGR:RunModel]{airGR::RunModel},
\link{RunModel.GRiwrmInputsModel}, \link{RunModel.GRiwrmOutputsModel}, \link{RunModel.Supervisor}}

\item{...}{Not used}
}
\value{
An object \strong{OutputsModel} with merged times series of simulation
results.
}
\description{
Merge Two outputs of airGR simulations
}
