% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OutputsModelQsim}
\alias{OutputsModelQsim}
\title{Create a data.frame with simulated flows at each nodes of the GRiwrm object}
\usage{
OutputsModelQsim(InputsModel, OutputsModel, IndPeriod_Run)
}
\arguments{
\item{InputsModel}{a \code{GRiwrmInputsModel} object created by \link{CreateInputsModel.GRiwrm}}

\item{OutputsModel}{a \code{GRiwrmOutputsModel} object created by \link{RunModel.GRiwrmInputsModel} or \link{RunModel.Supervisor}}

\item{IndPeriod_Run}{an \link{integer} vector (See \link[airGR:CreateRunOptions]{airGR::CreateRunOptions})}
}
\value{
a \link{data.frame} containing the simulated flows (in m3/time step) structured with the following columns:
\itemize{
\item 'DatesR' containing the timestamps of the time series
\item one column by node with the simulated flows
}
}
\description{
Create a data.frame with simulated flows at each nodes of the GRiwrm object
}
\details{
This function can only be called inside \link{RunModel.GRiwrmInputsModel} or \link{RunModel.Supervisor}
because it needs a \code{GRiwrmInputsModel} object internally modified by these functions
(\code{Qupstream} updated with simulated flows).
}
