% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GRiwrmOutputsModel.R
\name{plot.GRiwrmOutputsModel}
\alias{plot.GRiwrmOutputsModel}
\title{Function which creates screen plots giving an overview of the model outputs in the GRiwrm network}
\usage{
\method{plot}{GRiwrmOutputsModel}(x, Qobs = NULL, Vobs = NULL, unit = "m3/s", ...)
}
\arguments{
\item{x}{[object of class \emph{GRiwrmOutputsModel}] see \link{RunModel.GRiwrmInputsModel} for details}

\item{Qobs}{(optional) \link{matrix} time series of observed flows
(for the same time steps than simulated) (mm/time step) with one column
by hydrological model output named with the node ID (See \link{CreateGRiwrm} for details)}

\item{Vobs}{(optional) \link{matrix} time series of observed or targeted storage
(for the same time steps than simulated) [m3] with one column
by hydrological model output named with the node ID (See \link{CreateGRiwrm} for details)}

\item{unit}{(optional) \link{character} flows unit ("m3/s" or "mm")}

\item{...}{Further arguments for \link[airGR:plot.OutputsModel]{airGR::plot.OutputsModel} and \link{plot}}
}
\value{
\link{list} of plots.
}
\description{
Function which creates screen plots giving an overview of the model outputs in the GRiwrm network
}
\details{
For examples of use see topics \link{RunModel.GRiwrmInputsModel}, \link{RunModel_Reservoir},
and \link{RunModel.Supervisor}.
}
