% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GRiwrm.R
\name{plot.mermaid}
\alias{plot.mermaid}
\title{Plot a mermaid diagram}
\usage{
\method{plot}{mermaid}(x, add = FALSE)
}
\arguments{
\item{x}{\link{character} mermaid diagram dialect}

\item{add}{\link{logical} to add the diagram on the existing plot}
}
\value{
Nothing, used for side effect.
}
\description{
Plot a mermaid diagram
}
\examples{
s <- "flowchart LR
A -> B"
class(s) <- "mermaid"
plot(s)
}
