% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Qm3s.R
\name{plot.Qm3s}
\alias{plot.Qm3s}
\title{Plot a \code{Qm3s} object (time series of simulated flows)}
\usage{
\method{plot}{Qm3s}(
  x,
  type = "l",
  xlab = "Date",
  ylab = "Flow (m3/s)",
  main = "Simulated flows",
  col = rainbow(ncol(x) - 1),
  legend = colnames(x)[-1],
  legend.cex = 0.7,
  lty = 1,
  ...
)
}
\arguments{
\item{x}{a \link{data.frame} with a first column with \link{POSIXt} dates and followings columns with flows at each node of the network}

\item{type}{1-character string (See \link{plot.default}), default "l"}

\item{xlab}{a label for the x axis, defaults to "Date"}

\item{ylab}{a label for the y axis, defaults to "Flow (m3/s)"}

\item{main}{a main title for the plot, defaults to "Simulated flows"}

\item{col}{plotting color (See \link{par}), defaults to rainbow colors}

\item{legend}{See parameter \code{legend} of \link{legend}. Set to \link{NULL} to not display the legend}

\item{legend.cex}{\code{cex} parameter for the text of the legend (See \link{par})}

\item{lty}{The line type (See \link{par})}

\item{...}{Further arguments to pass to the \link{matplot} functions}
}
\description{
Plot a \code{Qm3s} object (time series of simulated flows)
}
