% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.R
\name{RunModel}
\alias{RunModel}
\title{RunModel function for both \strong{airGR} InputsModel and GRiwrmInputsModel object}
\usage{
RunModel(x, ...)
}
\arguments{
\item{x}{[object of class \emph{InputsModel} or \emph{GRiwrmInputsModel}] see \link{CreateInputsModel} for details}

\item{...}{further arguments passed to or from other methods}
}
\value{
Either a \link{list} of OutputsModel object (for GRiwrmInputsModel) or an OutputsModel object (for InputsModel)
}
\description{
RunModel function for both \strong{airGR} InputsModel and GRiwrmInputsModel object
}
