% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.SD.R
\name{RunModel.SD}
\alias{RunModel.SD}
\title{Run SD Model from run-off model outputs}
\usage{
\method{RunModel}{SD}(x, RunOptions, Param, QsimDown, ...)
}
\arguments{
\item{x}{\code{InputsModel} used as \code{InputsModel} parameter for \link[airGR:RunModel_Lag]{airGR::RunModel_Lag}}

\item{RunOptions}{[object of class \emph{RunOptions}] see \code{\link[airGR]{CreateRunOptions}} for details}

\item{Param}{[numeric] vector of 1 parameter
    \tabular{ll}{
      Velocity \tab Mean flow velocity [m/s]
    }
  }

\item{QsimDown}{a \link{numeric} corresponding to the runoff of the sub-basin (Typically the \code{Qsim} outputs of the GR model)}

\item{...}{further arguments passed to or from other methods}
}
\value{
\code{OutputsModel} object. See \link[airGR:RunModel_Lag]{airGR::RunModel_Lag}
}
\description{
Run SD Model from run-off model outputs
}
