% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setDataToLocation}
\alias{setDataToLocation}
\title{Write data to model input for the current time step}
\usage{
setDataToLocation(ctrlr, sv)
}
\arguments{
\item{ctrlr}{a \code{Controller} object (See \link{CreateController})}

\item{sv}{\code{Supervisor} (See \link{CreateSupervisor})}
}
\value{
\link{NULL}
}
\description{
Write data to model input for the current time step
}
