% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeRanking.R
\name{sort.GRiwrm}
\alias{sort.GRiwrm}
\title{Sort a GRiwrm network in upstream-downstream order ready for Calibration}
\usage{
\method{sort}{GRiwrm}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{A \emph{GRiwrm} object (See \link{CreateGRiwrm})}

\item{decreasing}{logical.  Should the sort be increasing or decreasing?
    Not available for partial sorting.}

\item{...}{arguments to be passed to or from methods or (for the
    default methods and objects without a class) to \code{sort.int}.}
}
\value{
The sorted \emph{GRiwrm} object in upstream-downstream order ready for Calibration
}
\description{
It Uses \link{getNodeRanking} for determining the best order for calibration and leaves
direct injection nodes at the tail of the list.
}
