% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.Calibration.R
\name{transferGRparams}
\alias{transferGRparams}
\title{Transfer GR parameters from one donor sub-basin to a receiver sub-basin}
\usage{
transferGRparams(
  InputsModel,
  Param,
  donor,
  receiver,
  default_param = NULL,
  verbose = FALSE
)
}
\arguments{
\item{InputsModel}{[\Object of class \emph{GRiwrmInputsModel}] see [CreateInputsModel.GRiwrm] for details

[\Object of class \emph{GRiwrmInputsModel}]: R:\%5C\%5CObject\%20of\%20class\%20\emph{GRiwrmInputsModel}\%5C
[CreateInputsModel.GRiwrm]: R:CreateInputsModel.GRiwrm}

\item{Param}{[\numeric] vector of GR model parameters

[\numeric]: R:\%5C\%5Cnumeric\%5C}

\item{donor}{[\character] id of the node which gives its parameters

[\character]: R:\%5C\%5Ccharacter\%5C}

\item{receiver}{[\character] id of the node which receives the parameters from the donor

[\character]: R:\%5C\%5Ccharacter\%5C}

\item{default_param}{[\numeric] vector of GR model parameters if parameters are missing from the donor

[\numeric]: R:\%5C\%5Cnumeric\%5C}

\item{verbose}{[\logical] Add information message on donor and receiver

[\logical]: R:\%5C\%5Clogical\%5C}
}
\value{
[\numeric] vector of transferred parameters

[\numeric]: R:\%5C\%5Cnumeric\%5C
}
\description{
This function is used by \code{Calibration.GRiwrmInputsModel} for transferring parameters
to ungauged nodes
}
\details{
\code{donor} and \code{receiver} nodes should have the same GR model with the same snow
module configuration.

The transfer takes care of:
\itemize{
\item the presence/absence of hydraulic routing parameters between the donor and the receiver
\item the transformation of the X4 parameters of GR models. Indeed, this parameter is correlated to the catchment area.
Therefore, it has to be rescaled from the surface of the donor catchment to the surface of the receiver catchment,
following the relationship suggested by Lobligeois (2014): \eqn{X4_{receiver} = X4_{donor} \times (S_{receiver} / S_{donor})^{0.3}}.
}
}
\references{
Lobligeois, F., 2014. Mieux connaitre la distribution spatiale des pluies améliore-t-il la modélisation des crues ?
Diagnostic sur 181 bassins versants français. Thèse de Doctorat, Irstea (Antony), AgroParisTech (Paris), 312 pp.
https://webgr.inrae.fr/Media/Files/biblio-theses/2014_lobligeois_these.pdf
}
