% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.GRiwrmInputsModel.R
\name{Calibration.GRiwrmInputsModel}
\alias{Calibration.GRiwrmInputsModel}
\title{Calibration of a semi-distributed precipitation-runoff model}
\usage{
\method{Calibration}{GRiwrmInputsModel}(
  InputsModel,
  RunOptions,
  InputsCrit,
  CalibOptions,
  useUpstreamQsim = TRUE,
  ...
)
}
\arguments{
\item{InputsModel}{object of class \emph{GRiwrmInputsModel}, see \link{CreateInputsModel.GRiwrm} for details}

\item{RunOptions}{object of class \emph{GRiwrmRunOptions}, see \link{CreateRunOptions.GRiwrmInputsModel} for details}

\item{InputsCrit}{object of class \emph{GRiwrmInputsCrit}, see \link{CreateInputsCrit.GRiwrmInputsModel} for details}

\item{CalibOptions}{object of class \emph{GRiwrmCalibOptions}, see \link{CreateCalibOptions.GRiwrmInputsModel} for details}

\item{useUpstreamQsim}{boolean describing if simulated (\code{TRUE}) or observed (\code{FALSE}) flows are used for calibration. Default is \code{TRUE}}

\item{...}{further arguments passed to \link[airGR:Calibration]{airGR::Calibration}.}
}
\value{
\link{list} of \link[airGR:Calibration]{airGR::Calibration} outputs for each node of the semi-distributed model.
}
\description{
Calibration of a semi-distributed precipitation-runoff model
}
