% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.GRiwrmInputsModel.R,
%   R/Calibration.InputsModel.R, R/Calibration.R
\name{Calibration.GRiwrmInputsModel}
\alias{Calibration.GRiwrmInputsModel}
\alias{Calibration.InputsModel}
\alias{Calibration}
\title{Calibration of the parameters of one catchment or a network of sub-catchments}
\usage{
\method{Calibration}{GRiwrmInputsModel}(
  InputsModel,
  RunOptions,
  InputsCrit,
  CalibOptions,
  useUpstreamQsim = TRUE,
  ...
)

\method{Calibration}{InputsModel}(InputsModel, CalibOptions, ...)

Calibration(InputsModel, ...)
}
\arguments{
\item{InputsModel}{[object of class \emph{InputsModel} or \emph{GRiwrmInputsModel}] see \link{CreateInputsModel}}

\item{RunOptions}{[object of class \emph{RunOptions} or \emph{GRiwrmRunOptions}] see \link{CreateRunOptions}}

\item{InputsCrit}{[object of class \emph{InputsCrit} or \emph{GRiwrmInputsCrit}] see \link{CreateInputsCrit}}

\item{CalibOptions}{[object of class \emph{CalibOptions} or \emph{GRiwrmCalibOptions}] see \link{CreateCalibOptions} for details}

\item{useUpstreamQsim}{boolean describing if simulated (\code{TRUE}) or observed (\code{FALSE}) flows are used for calibration. Default is \code{TRUE}}

\item{...}{further arguments passed to \link[airGR:Calibration]{airGR::Calibration}, see details}
}
\value{
Depending on the class of \code{InputsModel} argument (respectively
\code{InputsModel} or \code{GRiwrmInputsModel} object), the returned value is respectively:
\itemize{
\item an \code{OutputsCalib} object (See \link[airGR:Calibration]{airGR::Calibration} for more details on this object)
\item a \code{GRiwrmOutputsCalib} object which is a \link{list} of \code{OutputsCalib} objects with
one item per modeled sub-catchment
}
}
\description{
Calibration algorithm that optimizes the error criterion selected as objective
function using the provided functions.
}
\details{
This function can be used either for a catchment (with an \emph{InputsModel}
object), for a network (with a \emph{GRiwrmInputsModel} object), or for an
ungauged node cluster (with a \emph{Ungauged} object).

Argument classes should be consistent to the usage:
\itemize{
\item a \code{InputsModel} argument of class \emph{InputsModel} must be followed by a
\code{RunOptions} argument of class \emph{RunOptions}, an \code{InputsCrit} argument of
class \emph{InputsCrit} and a \code{CalibOptions} of class \emph{CalibOptions}
\item an \code{InputsModel} argument of class \emph{GRiwrmInputsModel} must be followed
by a \code{RunOptions} argument of class \emph{GRiwrmRunOptions}, an \code{InputsCrit}
argument of class \emph{GRiwrmInputsCrit} and a \code{CalibOptions} of class
\emph{GRiwrmCalibOptions}
}
}
\examples{
# Loading catchment data
data(Severn)

#############################################################################
# EXAMPLE 1 - Basic calibration workflow                                   #
#############################################################################

# Creating catchment network
nodes <- Severn$BasinsInfo[, c("gauge_id", "downstream_id",
                               "distance_downstream", "area")]
nodes$model <- "RunModel_GR4J"
rename_columns <- list(id = "gauge_id",
                       down = "downstream_id",
                       length = "distance_downstream")
griwrm <- CreateGRiwrm(nodes, rename_columns)
griwrm

# Preparation of InputsModel object
BasinsObs <- Severn$BasinsObs
DatesR <- BasinsObs[[1]]$DatesR
PrecipTot <- cbind(sapply(BasinsObs, function(x) {x$precipitation}))
PotEvapTot <- cbind(sapply(BasinsObs, function(x) {x$peti}))
Qobs <- cbind(sapply(BasinsObs, function(x) {x$discharge_spec}))
Precip <- ConvertMeteoSD(griwrm, PrecipTot)
PotEvap <- ConvertMeteoSD(griwrm, PotEvapTot)
InputsModel <- CreateInputsModel(griwrm, DatesR, Precip, PotEvap)

# Calibration period selection
# Set aside warm-up period and use the rest for calibration
IndPeriod_Run <- seq(
  which(InputsModel[[1]]$DatesR ==
        (InputsModel[[1]]$DatesR[1] + 365 * 24 * 60 * 60)),
  length(InputsModel[[1]]$DatesR)
)
IndPeriod_WarmUp <- seq(1, IndPeriod_Run[1] - 1)

# Preparation of RunOptions object
RunOptions <- CreateRunOptions(
  InputsModel,
  IndPeriod_WarmUp = IndPeriod_WarmUp,
  IndPeriod_Run = IndPeriod_Run
)

# Calibration criterion: preparation of the InputsCrit object
InputsCrit <- CreateInputsCrit(
  InputsModel = InputsModel,
  FUN_CRIT = ErrorCrit_KGE2,
  RunOptions = RunOptions,
  Obs = Qobs[IndPeriod_Run, ],
  transfo = "sqrt"
)

# Preparation of CalibOptions object
CalibOptions <- CreateCalibOptions(InputsModel)

if (interactive()) { # Too long for CRAN check...
  # Calibration
  OutputsCalib <- suppressWarnings(
    Calibration(InputsModel, RunOptions, InputsCrit, CalibOptions)
  )

  # Simulation
  OutputsModels <- RunModel(
    InputsModel,
    RunOptions = RunOptions,
    Param = extractParam(OutputsCalib)
  )
}

}
\seealso{
\code{\link[=CreateGRiwrm]{CreateGRiwrm()}}, \code{\link[=CreateInputsModel.GRiwrm]{CreateInputsModel.GRiwrm()}}, \code{\link[=CreateInputsCrit]{CreateInputsCrit()}}, \code{\link[=CreateCalibOptions]{CreateCalibOptions()}}
}
