% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{Calibration}
\alias{Calibration}
\title{Calibration of either a \strong{airGR} model or a \strong{airGRiwrm} semi-distributed model}
\usage{
Calibration(InputsModel, ...)
}
\arguments{
\item{InputsModel}{the class of the first parameter determines which calibration algorithm is used}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\emph{OutputsCalib} or \emph{GRiwrmOutputsCalib} object
}
\description{
Calibration of either a \strong{airGR} model or a \strong{airGRiwrm} semi-distributed model
}
