% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration.R
\name{Calibration}
\alias{Calibration}
\title{Calibration of either \strong{airGR} model and \strong{airGRiwrm} semi-distributive model}
\usage{
Calibration(InputsModel, ...)
}
\arguments{
\item{InputsModel}{the class of the first parameter determine which calibration is used}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\emph{OutputsCalib} or \emph{GRiwrmOutputsCalib} object
}
\description{
Calibration of either \strong{airGR} model and \strong{airGRiwrm} semi-distributive model
}
