% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkRunModelParameters}
\alias{checkRunModelParameters}
\title{Check of the parameters of RunModel methods}
\usage{
checkRunModelParameters(InputsModel, RunOptions, Param)
}
\arguments{
\item{InputsModel}{[\code{GRiwrmInputsModel} object] see \link{CreateInputsModel.GRiwrm} for details}

\item{RunOptions}{[\code{GRiwrmRunOptions} object] see \link{CreateRunOptions.GRiwrmInputsModel} for details}

\item{Param}{\link{list} of containing model parameter values of each node of the network}
}
\description{
Stop the execution if an error is detected.
}
