% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCalibOptions.GRiwrmInputsModel.R
\name{CreateCalibOptions.GRiwrmInputsModel}
\alias{CreateCalibOptions.GRiwrmInputsModel}
\title{Title}
\usage{
\method{CreateCalibOptions}{GRiwrmInputsModel}(InputsModel, ...)
}
\arguments{
\item{InputsModel}{object of class \emph{GRiwrmInputsModel}, see \link{CreateInputsModel.GRiwrm} for details.}

\item{...}{further arguments passed to \link[airGR:CreateCalibOptions]{airGR::CreateCalibOptions}.}
}
\value{
\emph{GRiwrmCalibOptions} object.
}
\description{
Title
}
