% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCalibOptions.GRiwrmInputsModel.R,
%   R/CreateCalibOptions.InputsModel.R, R/CreateCalibOptions.R
\name{CreateCalibOptions.GRiwrmInputsModel}
\alias{CreateCalibOptions.GRiwrmInputsModel}
\alias{CreateCalibOptions.InputsModel}
\alias{CreateCalibOptions}
\title{Creation of the CalibOptions object}
\usage{
\method{CreateCalibOptions}{GRiwrmInputsModel}(InputsModel, ...)

\method{CreateCalibOptions}{InputsModel}(InputsModel, ...)

CreateCalibOptions(InputsModel, ...)
}
\arguments{
\item{InputsModel}{object of class \emph{InputsModel} or \emph{GRwirmInputsModel}. See \link{CreateInputsModel} for details}

\item{...}{arguments passed to \link[airGR:CreateCalibOptions]{airGR::CreateCalibOptions}, see details}
}
\value{
Depending on the class of \code{InputsModel} argument (respectively \code{InputsModel} and \code{GRiwrmInputsModel} object), the returned value is respectively:
\itemize{
\item a \code{CalibOptions} object (See \link[airGR:CreateCalibOptions]{airGR::CreateCalibOptions})
\item a \code{GRiwrmCalibOptions} object which is a \link{list} of \code{CalibOptions} object with one item per modelled sub-catchment
}
}
\description{
This function can be used either for a catchment (with an \emph{InputsModel} object) or for a network (with a \emph{GRiwrmInputsModel} object)
}
\details{
See \link[airGR:CreateCalibOptions]{airGR::CreateCalibOptions} documentation for a complete list of arguments.

With a \emph{GRwirmInputsModel} object, all arguments are applied on each sub-catchments of the network.
}
