% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCalibOptions.R
\name{CreateCalibOptions}
\alias{CreateCalibOptions}
\title{Creation of the CalibOptions object either for \emph{InputsModel} or for \emph{GRwirmInputsModel} objects}
\usage{
CreateCalibOptions(InputsModel, ...)
}
\arguments{
\item{InputsModel}{object of class \emph{InputsModel} or \emph{GRwirmInputsModel}. See \link{CreateInputsModel} for details}

\item{...}{further arguments passed to or from other methods}
}
\value{
\link{list} Either a \code{CalibOptions} or a \code{GRiwrmCalibOptions} object
}
\description{
Creation of the CalibOptions object either for \emph{InputsModel} or for \emph{GRwirmInputsModel} objects
}
