% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createController.R
\name{createControl}
\alias{createControl}
\title{Creation of a list of controls for command (U) and controlled variables (Y)}
\usage{
createControl(locations)
}
\arguments{
\item{locations}{\link{character} containing the location of the variable in the model (see details)}
}
\value{
\link{matrix} with each column being the location of the variables and the rows being
the values of the variable for the current time steps (empty by default)
}
\description{
Creation of a list of controls for command (U) and controlled variables (Y)
}
\examples{
# For pointing the discharge at the oulet of basins "54095" and "54002"
createControl(c("54095", "54002"))
}
