% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGRiwrm.R
\name{CreateGRiwrm}
\alias{CreateGRiwrm}
\alias{GRiwrm}
\title{Generation of a network description containing all hydraulic nodes and the
description of their connections}
\usage{
CreateGRiwrm(
  db,
  cols = list(id = "id", down = "down", length = "length", model = "model", area =
    "area"),
  keep_all = FALSE
)
}
\arguments{
\item{db}{\link{data.frame} description of the network (See details)}

\item{cols}{\link{list} or \link{vector} columns of \code{db}. By default, mandatory column
names are: \code{id}, \code{down}, \code{length}, \code{area} and \code{model}. Other names can be
handled with a named list or vector containing items defined as \code{"required name" = "column name in db"} (See details)}

\item{keep_all}{\link{logical} indicating if all columns of \code{db} should be kept
or if only columns defined in \code{cols} should be kept}
}
\value{
\link{data.frame} of class \code{GRiwrm} describing the airGR semi-distributed
model network, with each line corresponding to a location on the river
network and with the following columns:
\itemize{
\item \code{id} (\link{character}): node identifier
\item \code{down} (\link{character}): identifier of the node downstream of the current
node (\link{NA} for the most downstream node)
\item \code{length} (\link{numeric}): hydraulic distance to the downstream node in km
(\link{NA} for the most downstream node)
\item \code{area} (\link{numeric}): total area of the basin starting from the current
node location in km2
\item \code{model} (\link{character}): hydrological model to use (\link{NA} for using observed
flow instead of a runoff model output)
\item \code{donor} (\link{character}): node used as "donor" for the the model and the
calibration parameters.
}
}
\description{
Generation of a network description containing all hydraulic nodes and the
description of their connections
}
\details{
\code{db} is a \link{data.frame} which at least contains in its columns:
\itemize{
\item a node identifier (column \code{id}),
\item the identifier and the hydraulic distance to the downstream node
(\link{character} columns \code{down} and \link{numeric} columns \code{length} in km). The
last downstream node should have fields \code{down} and \code{length} set to \code{NA},
\item the total area of the basin at the node location (\link{numeric} column \code{area} in km2).
Direct injection node can have a null area defined by \code{NA}
\item the model to use (\link{character} column \code{model}), see section below for details
}
\subsection{Available models in airGRiwrm}{

The "model" column should be filled by one of the following:
\itemize{
\item One of the hydrological models available in the \emph{airGR} package defined by its
\code{RunModel} function (i.e.: \code{RunModel_GR4J}, \code{RunModel_GR5HCemaneige}...)
\item \code{Ungauged} for an ungauged node. The sub-basin inherits hydrological model and
parameters from a "donor" sub-basin. By default the donor is the first gauged
node at downstream
\item \code{NA} for injecting (or abstracting) a flow time series at the location of the node
(direct flow injection)
\item \code{Diversion} for abstracting a flow time series from an existing node transfer it
to another node. As a \code{Diversion} is attached to an existing node, this node is
then described with 2 lines: one for the hydrological model and another one for the
diversion
}
}
}
\examples{
library(airGRiwrm)

#########################################
# Network of 2 nodes distant of 150 km: #
#########################################
# - an upstream reservoir modelled as a direct flow injection (no model)
# - a gauging station downstream a catchment of 360 km² modelled with GR4J
db <- data.frame(id = c("Reservoir", "GaugingDown"),
                 length = c(150, NA),
                 down = c("GaugingDown", NA),
                 area = c(NA, 360),
                 model = c(NA, "RunModel_GR4J"),
                 stringsAsFactors = FALSE)
griwrm_basic <- CreateGRiwrm(db)
griwrm_basic
# Network diagram with direct flow node in red, intermediate sub-basin in green
plot(griwrm_basic)

###################################################
# GR4J semi-distributed model of the Severn River #
###################################################
data(Severn)
nodes <- Severn$BasinsInfo
nodes$model <- "RunModel_GR4J"
str(nodes)
# Mismatch column names are renamed to stick with GRiwrm requirements
rename_columns <- list(id = "gauge_id",
                       down = "downstream_id",
                       length = "distance_downstream")
griwrm_severn <- CreateGRiwrm(nodes, rename_columns)
griwrm_severn
# Network diagram with upstream basin nodes in blue, intermediate sub-basin in green
plot(griwrm_severn)

####################################################################
# Severn network with an ungauged station at nodes 54029 and 54001 #
####################################################################
# By default the first gauged node at downstream is used for parameter calibration (54032)
nodes_ungauged <- nodes
nodes_ungauged$model[nodes_ungauged$gauge_id \%in\% c("54029", "54001")] <- "Ungauged"
griwrm_ungauged <- CreateGRiwrm(nodes_ungauged, rename_columns)
# The `donor` column define which node is used for parameter calibration
griwrm_ungauged
# Network diagram with gauged nodes of vivid color, and ungauged nodes of dull color
plot(griwrm_ungauged)

#######################################################
# Severn network with a Diversion on the node "54029" #
# which transfer flows to the node "54001"            #
#######################################################
nodes_div <- nodes[, c("gauge_id", "downstream_id", "distance_downstream", "model", "area")]
nodes_div <- rbind(nodes_div, data.frame(gauge_id = "54029",
                                         downstream_id = "54001",
                                         distance_downstream = 20,
                                         model = "Diversion",
                                         area = NA))
griwrm_div <- CreateGRiwrm(nodes_div, rename_columns)
# Network diagram figures Diversion node by a red frame and a red arrow
plot(griwrm_div)
}
