% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInputsCrit.GRiwrmInputsModel.R
\name{CreateInputsCrit.GRiwrmInputsModel}
\alias{CreateInputsCrit.GRiwrmInputsModel}
\title{Create \emph{GRiwrmInputsCrit} object for \strong{airGRiwrm}.}
\usage{
\method{CreateInputsCrit}{GRiwrmInputsModel}(
  InputsModel,
  FUN_CRIT = airGR::ErrorCrit_NSE,
  RunOptions,
  Obs,
  ...
)
}
\arguments{
\item{InputsModel}{object of class \emph{GRiwrmInputsModel}, see \link{CreateInputsModel.GRiwrm} for details.}

\item{FUN_CRIT}{[function (atomic or list)] error criterion function (e.g. \link[airGR:ErrorCrit_RMSE]{airGR::ErrorCrit_RMSE}, \link[airGR:ErrorCrit_NSE]{airGR::ErrorCrit_NSE})}

\item{RunOptions}{object of class \emph{GRiwrmRunOptions}, see \link{CreateRunOptions.GRiwrm} for details.}

\item{Obs}{matrix or data frame containing observed flows. Column names correspond to nodes ID}

\item{...}{further arguments passed to \link[airGR:CreateInputsCrit]{airGR::CreateInputsCrit}.}
}
\value{
Object of class \emph{GRiwrmInputsCrit} which is a list of \code{airGR::InputsCrit} objects (See \link[airGR:CreateInputsCrit]{airGR::CreateInputsCrit})
}
\description{
This function does the same operations as \link[airGR:CreateInputsCrit]{airGR::CreateInputsCrit} for all sub-basins of the GRiwrm model.
}
