% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInputsCrit.R
\name{CreateInputsCrit}
\alias{CreateInputsCrit}
\title{Creation of the InputsCrit object required to the ErrorCrit functions}
\usage{
CreateInputsCrit(InputsModel, ...)
}
\arguments{
\item{InputsModel}{InputsModel for \strong{airGRiwrm} (See \code{[CreateInputsModel.GRiwrm]}) or \strong{airGR} (See \code{\link[airGR]{CreateInputsModel}})}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Either a \code{InputsCrit} or a \code{GRiwrmInputsCrit} object
}
\description{
Creation of the InputsCrit object required to the ErrorCrit functions
}
