% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInputsModel.default.R
\name{CreateInputsModel.default}
\alias{CreateInputsModel.default}
\title{Wrapper for \code{\link[airGR]{CreateInputsModel}} for one sub-basin.}
\usage{
\method{CreateInputsModel}{default}(x, ...)
}
\arguments{
\item{x}{\link{function} hydrological model function (e.g. \code{\link[airGR]{RunModel_GR4J}}...)}

\item{...}{arguments passed to \code{\link[airGR]{CreateInputsModel}}}
}
\description{
Wrapper for \code{\link[airGR]{CreateInputsModel}} for one sub-basin.
}
