% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateInputsModel.R
\name{CreateInputsModel}
\alias{CreateInputsModel}
\alias{CreateInputsModel.default}
\title{Generic function for creating \code{InputsModel} object for either \strong{airGR} or \strong{airGRiwrm}}
\usage{
CreateInputsModel(x, ...)

\method{CreateInputsModel}{default}(x, ...)
}
\arguments{
\item{x}{First parameter determining which InputsModel object is created}

\item{...}{further arguments passed to or from other methods.}
}
\value{
InputsModel or GRiwrmInputsObject object
}
\description{
See the methods \link{CreateInputsModel.GRiwrm} for \strong{airGRiwrm} and \link[airGR:CreateInputsModel]{airGR::CreateInputsModel} for \strong{airGR}.
}
\seealso{
\code{\link[=CreateInputsModel.GRiwrm]{CreateInputsModel.GRiwrm()}}, \code{\link[airGR:CreateInputsModel]{airGR::CreateInputsModel()}}
}
