% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRunOptions.InputsModel.R
\name{CreateRunOptions.InputsModel}
\alias{CreateRunOptions.InputsModel}
\title{Wrapper for \code{\link[airGR]{CreateRunOptions}} for one sub-basin.}
\usage{
\method{CreateRunOptions}{InputsModel}(InputsModel, ...)
}
\arguments{
\item{InputsModel}{object of class \emph{InputsModel}, see \code{\link[airGR]{CreateInputsModel}} for details.}

\item{...}{Arguments passed to \code{\link[airGR]{CreateRunOptions}}}
}
\description{
Wrapper for \code{\link[airGR]{CreateRunOptions}} for one sub-basin.
}
