% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRunOptions.R
\name{CreateRunOptions}
\alias{CreateRunOptions}
\title{Create \emph{RunOptions} object for \strong{airGR} and \strong{airGRiwrm}. See \code{\link[airGR]{CreateRunOptions}} and \code{[CreateOptions.GRiwrmInputsModel]}.}
\usage{
CreateRunOptions(InputsModel, ...)
}
\arguments{
\item{InputsModel}{object of class \emph{InputsModel} (see \code{\link[airGR]{CreateInputsModel}}) or \emph{GRiwrmInputsModel} (See \code{[CreateInputsModel.GRiwrm]}).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Object of \emph{RunOptions} class family
}
\description{
Create \emph{RunOptions} object for \strong{airGR} and \strong{airGRiwrm}. See \code{\link[airGR]{CreateRunOptions}} and \code{[CreateOptions.GRiwrmInputsModel]}.
}
