% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.Calibration.R
\name{extractParam}
\alias{extractParam}
\alias{extractParam.GRiwrmOutputsCalib}
\alias{extractParam.GRiwrmOutputsModel}
\title{Extract calibrated parameters}
\usage{
extractParam(x)

\method{extractParam}{GRiwrmOutputsCalib}(x)

\method{extractParam}{GRiwrmOutputsModel}(x)
}
\arguments{
\item{x}{A \emph{GRiwrmOutputsModel} object returned by \link{Calibration.GRiwrmInputsModel}}
}
\value{
A named \link{list} of \link{numeric} \link{vector} containing the calibrated parameters
of each modeled node.
}
\description{
Extract \link{list} of parameters from the output of \link{Calibration.GRiwrmInputsModel}
which can be directly used as argument \code{Param} of \link{RunModel.GRiwrmInputsModel}
and \link{RunModel.Supervisor}.
}
\details{
See vignettes and example of \link{RunModel_Reservoir} for examples of use.
}
\seealso{
\link{Calibration}, \link{RunModel.GRiwrmInputsModel}, \link{RunModel.Supervisor}
}
