% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDataFromLocation}
\alias{getDataFromLocation}
\title{Retrieval of data in the model for the current time steps}
\usage{
getDataFromLocation(loc, sv)
}
\arguments{
\item{loc}{location of the data}

\item{sv}{[object of class \code{Supervisor}] see \link{CreateSupervisor} for details}
}
\value{
\link{numeric} retrieved data at the location
}
\description{
Function to be called inside a Supervisor
}
