% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.RunModel.R
\name{getNextTimeSteps}
\alias{getNextTimeSteps}
\title{Get the next time steps date/time of a simulation}
\usage{
getNextTimeSteps(x, TimeStep = 1L)
}
\arguments{
\item{x}{Object returned by \link{RunModel.GRiwrmInputsModel},
\link{RunModel.Supervisor}, or \link{RunModel.GRiwrmOutputsModel}}

\item{TimeStep}{\link{integer} number of time steps to get after the end of the
simulation}
}
\value{
A \link{POSIXct} containing the date/time of the time steps following
the end of the simulation.
}
\description{
Get the next time steps date/time of a simulation
}
