% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeProperties.R
\name{getNodeProperties}
\alias{getNodeProperties}
\title{Get list of properties of a GRiwrm network node}
\usage{
getNodeProperties(id, griwrm)
}
\arguments{
\item{id}{\link{character} Id of the node in the GRiwrm object}

\item{griwrm}{[GRiwrm object] describing the network of the semi-distributed model (See \link{CreateGRiwrm})}
}
\value{
A \link{list} with the following items:
\itemize{
\item "position" (\link{character}): Position of the node in the network ("Upstream" or "Intermediate")
\item "hydrology" (\link{character}): describe if the node is a "Gauged" or an "Ungauged" station
modelled with an hydrological model, or a "DirectionInjection" node
\item "Upstream" (\link{logical}): is the node an upstream node?
\item "DirectInjection" (\link{logical}): is the node a Direct Injection node?
\item "Diversion" (\link{logical}): is the node a Diversion node?
}
}
\description{
Get list of properties of a GRiwrm network node
}
\examples{
###############################################################################
# Severn network with :                                                       #
# - a Diversion on the node "54001" which transfer flows to the node "540029" #
# - node 54002 as a Direct Injection node                                     #
###############################################################################
data(Severn)
nodes <- Severn$BasinsInfo
nodes$model <- "RunModel_GR4J"
str(nodes)
nodes <- nodes[, c("gauge_id", "downstream_id", "distance_downstream", "model", "area")]
# Add a Diversion node from node "54001" to "54029"
nodes <- rbind(nodes,
               data.frame(
                 gauge_id = "54001",
                 downstream_id = "54029",
                 distance_downstream = 20,
                 model = "Diversion",
                 area = NA
               ))
# Set node '54002' as a Direct Injection node
nodes$model[nodes$id == "54002"] <- NA
# Mismatch column names are renamed to stick with GRiwrm requirements
rename_columns <- list(id = "gauge_id",
                       down = "downstream_id",
                       length = "distance_downstream")
# Create GRiwrm object and display properties
griwrm <- CreateGRiwrm(nodes, rename_columns)

str(getNodeProperties("54001", griwrm))

getAllNodesProperties(griwrm)
}
