% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateGRiwrm.R
\name{getNodeRanking}
\alias{getNodeRanking}
\title{Sort the nodes from upstream to downstream.}
\usage{
getNodeRanking(griwrm)
}
\arguments{
\item{griwrm}{See \link{CreateGRiwrm}}
}
\value{
vector with the ordered node names.
}
\description{
Sort the nodes from upstream to downstream.
}
