% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodeRanking.R
\name{getNodeRanking}
\alias{getNodeRanking}
\title{Sorting of the nodes from upstream to downstream for RunModel and Calibration}
\usage{
getNodeRanking(griwrm)
}
\arguments{
\item{griwrm}{[object of class \code{GRiwrm}] see \link{CreateGRiwrm} for details}
}
\value{
A \link{character} \link{vector} containing ordered node ids
}
\description{
Sorting of the nodes from upstream to downstream for RunModel and Calibration
}
\seealso{
\code{\link[=CreateGRiwrm]{CreateGRiwrm()}}
}
