% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSD_Ids}
\alias{getSD_Ids}
\title{Function to obtain the ID of sub-basins using SD model}
\usage{
getSD_Ids(InputsModel)
}
\arguments{
\item{InputsModel}{[\code{GRiwrmInputsModel} object]}
}
\value{
\link{character} IDs of the sub-basins using SD model
}
\description{
Function to obtain the ID of sub-basins using SD model
}
