% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRiwrm.R
\name{GRiwrm}
\alias{GRiwrm}
\title{Generate a network description containing all hydraulic nodes and the description
of their connections}
\usage{
GRiwrm(
  db,
  cols = list(id = "id", down = "down", length = "length", model = "model", area =
    "area"),
  keep_all = FALSE
)
}
\arguments{
\item{db}{a \link{data.frame} containing the description of the network (See details)}

\item{cols}{named list or vector for matching columns of \code{db} parameter. By default, mandatory columns names are: \code{id}, \code{down}, \code{length}. But other names can be handled with a named list or vector containing items defined as \code{"required name" = "column name in db"}}

\item{keep_all}{keep all column of \code{db} or keep only columns defined in \code{cols}}
}
\value{
\code{GRiwrm} class object containing the description of the semi-distributed model network
}
\description{
Generate a network description containing all hydraulic nodes and the description
of their connections
}
\details{
\code{db} is a \link{data.frame} which at least contains in its columns:
\itemize{
\item a node identifier (column \code{id}),
\item the identifier and the hydraulic distance to the downstream node (\link{character} columns \code{down} and \link{numeric} columns \code{length} in meters). The last downstream node should have fields \code{down} and \code{length} set to \code{NA},
\item the area of the basin (\link{numeric} column \code{area} in km^2^)
\item the hydrological model to use if so (\link{character} column \code{model}).
}
}
\examples{
#################################################################
# Run the `airGRRunModel_Lag` example in the GRiwrm fashion way #
#################################################################

# Run airGR RunModel_Lag example for harvesting necessary data
library(airGR)
example(RunModel_Lag)
# detach the package because airGR overwrite airGRiwrm functions here
detach("package:airGR")

# This example is a network of 2 nodes which can be describe like this:
db <- data.frame(id = c("Reservoir", "GaugingDown"),
                 length = c(LengthHydro, NA),
                 down = c("GaugingDown", NA),
                 area = c(NA, BasinInfo$BasinArea),
                 model = c(NA, "RunModel_GR4J"),
                 stringsAsFactors = FALSE)

# Create GRiwrm object from the data.frame
griwrm <- GRiwrm(db)
str(griwrm)

}
