% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isNodeDownstream}
\alias{isNodeDownstream}
\title{Check if a node is downstream another one}
\usage{
isNodeDownstream(InputsModel, current_node, down_node)
}
\arguments{
\item{InputsModel}{[\code{GRiwrmInputsModel} object] see \link{CreateInputsModel.GRiwrm} for details}

\item{current_node}{\link{character} with the id of the current node}

\item{down_node}{\link{character} with the id of the node for which we want to know if it is downstream \code{current_node}}
}
\value{
\link{logical} \code{TRUE} if the node with the id \code{down_node} is downstream the node with the id \code{current_node}
}
\description{
Check if a node is downstream another one
}
