% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.GRiwrm.R
\name{isNodeDownstream}
\alias{isNodeDownstream}
\alias{isNodeDownstream.GRiwrmInputsModel}
\alias{isNodeDownstream.GRiwrm}
\alias{isNodeUpstream}
\alias{isNodeUpstream.GRiwrm}
\alias{isNodeUpstream.GRiwrmInputsModel}
\title{Check if a node is downstream or upstream another one}
\usage{
isNodeDownstream(x, current_node, candidate_node)

\method{isNodeDownstream}{GRiwrmInputsModel}(x, current_node, candidate_node)

\method{isNodeDownstream}{GRiwrm}(x, current_node, candidate_node)

isNodeUpstream(x, current_node, candidate_node)

\method{isNodeUpstream}{GRiwrm}(x, current_node, candidate_node)

\method{isNodeUpstream}{GRiwrmInputsModel}(x, current_node, candidate_node)
}
\arguments{
\item{x}{[\code{GRiwrmInputsModel} object] (see \link{CreateInputsModel.GRiwrm}) or
[\code{GRiwrm} object] (See \link{CreateGRiwrm})}

\item{current_node}{\link{character} with the id of the current node}

\item{candidate_node}{\link{character} with the id of the node for which we want
to know if it is downstream or upstream \code{current_node}}
}
\value{
\link{logical} \code{TRUE} if the node with the id \code{down_candidate} is downstream
or upstream the node with the id \code{current_node}
}
\description{
Check if a node is downstream or upstream another one
}
