% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OutputsModelQsim}
\alias{OutputsModelQsim}
\title{Creation of a data.frame with simulated flows at each node of the GRiwrm object}
\usage{
OutputsModelQsim(InputsModel, OutputsModel, IndPeriod_Run)
}
\arguments{
\item{InputsModel}{[\code{GRiwrmInputsModel} object] see \link{CreateInputsModel.GRiwrm} for details}

\item{OutputsModel}{[\code{GRiwrmOutputsModel} object] see \link{RunModel.GRiwrmInputsModel} or \link{RunModel.Supervisor} for details}

\item{IndPeriod_Run}{\link{numeric} index of period to be used for the model run \link{-}. See \link[airGR:CreateRunOptions]{airGR::CreateRunOptions} for details}
}
\value{
a \link{data.frame} containing the simulated flows (in m3/time step) structured with the following columns:
\itemize{
\item 'DatesR' vector of dates  of the time series
\item one column by node with the simulated flows
}
}
\description{
Creation of a data.frame with simulated flows at each node of the GRiwrm object
}
\details{
This function can only be called inside \link{RunModel.GRiwrmInputsModel} or \link{RunModel.Supervisor}
because it needs a \code{GRiwrmInputsModel} object internally modified by these functions
(\code{Qupstream} updated with simulated flows).
}
