% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GRiwrmOutputsModel.R
\name{plot.GRiwrmOutputsModel}
\alias{plot.GRiwrmOutputsModel}
\title{Function which creates screen plots giving an overview of the models outputs in the GRiwrm network.}
\usage{
\method{plot}{GRiwrmOutputsModel}(x, Qobs = NULL, ...)
}
\arguments{
\item{x}{object of class \emph{GRiwrmOutputsModel}, see \link{RunModel.GRiwrmInputsModel}}

\item{Qobs}{(optional) \link{numeric} \link{matrix} time series of observed flow
(for the same time steps than simulated) (mm/time step) with one column
by hydrological model output named with the node Id (See \link{GRiwrm})}

\item{...}{Further arguments for \link[airGR:plot.OutputsModel]{airGR::plot.OutputsModel} and \link{plot}}
}
\value{
\link{list} of plots
}
\description{
Function which creates screen plots giving an overview of the models outputs in the GRiwrm network.
}
\examples{
#################################################################
# Run the `airGRRunModel_Lag` example in the GRiwrm fashion way #
#################################################################

# Run airGR RunModel_Lag example for harvesting necessary data
library(airGR)
example(RunModel_Lag)
# detach the package because airGR overwrite airGRiwrm functions here
detach("package:airGR")

# This example is a network of 2 nodes which can be describe like this:
db <- data.frame(id = c("Reservoir", "GaugingDown"),
                 length = c(LengthHydro, NA),
                 down = c("GaugingDown", NA),
                 area = c(NA, BasinInfo$BasinArea),
                 model = c(NA, "RunModel_GR4J"))

# Create GRiwrm object from the data.frame
griwrm <- GRiwrm(db)
str(griwrm)

# Formatting observations for the hydrological models
# Each input data should be a matrix or a data.frame with the good id in the name of the column
Precip <- matrix(BasinObs$P, ncol = 1)
colnames(Precip) <- "GaugingDown"
PotEvap <- matrix(BasinObs$E, ncol = 1)
colnames(PotEvap) <- "GaugingDown"

# Observed flows are integrated now because we mix:
#  - flows that are directly injected in the model
#  - flows that could be used for the calibration of the hydrological models
Qobs = matrix(c(Qupstream, BasinObs$Qmm), ncol = 2)
colnames(Qobs) <- griwrm$id
str(Qobs)

# Creation of the GRiwrmInputsModel object (= a named list of InputsModel objects)
InputsModels <- CreateInputsModel(griwrm,
                            DatesR = BasinObs$DatesR,
                            Precip = Precip,
                            PotEvap = PotEvap,
                            Qobs = Qobs)
str(InputsModels)

# Creation of the GriwmRunOptions object
RunOptions2 <- CreateRunOptions(InputsModel2,
                                IndPeriod_Run = Ind_Run)
str(RunOptions2)

# Parameters of the SD models should be encapsulated in a named list
Param2 <- list(`GaugingDown` = c(Velocity, Param))

# RunModel for the whole network
OutputsModels <- RunModel(InputsModels,
                          RunOptions = RunOptions2,
                          Param = Param2)
str(OutputsModels)

# Comparison between GRiwrm simulation and airGR simulation
plot(OutputsModels, Qobs = data.frame(`GaugingDown` = OutputsModel$Qsim))
}
