% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.GRiwrmInputsModel.R
\name{RunModel.GRiwrmInputsModel}
\alias{RunModel.GRiwrmInputsModel}
\title{RunModel function for GRiwrmInputsModel object}
\usage{
\method{RunModel}{GRiwrmInputsModel}(x, RunOptions, Param, ...)
}
\arguments{
\item{x}{object of class \emph{GRiwrmInputsModel}, see \code{[CreateInputsModel.GRiwrm]} for details.}

\item{RunOptions}{object of class \emph{GRiwrmRunOptions}, see \code{[CreateRunOptions.GRiwrm]} for details.}

\item{Param}{list of parameter. The list item names are the IDs of the sub-basins. Each item is a vector of numerical parameters.}

\item{...}{Mandatory for S3 method signature function compatibility with generic.}
}
\value{
\emph{GRiwrmOutputsModel} object which is a list of \emph{OutputsModel} objects (See \code{\link[airGR]{RunModel}}) for each node of the semi-distributed model.
}
\description{
RunModel function for GRiwrmInputsModel object
}
