% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.GRiwrmOutputsModel.R
\name{RunModel.GRiwrmOutputsModel}
\alias{RunModel.GRiwrmOutputsModel}
\title{RunModel for hot restart after a previous simulation period}
\usage{
\method{RunModel}{GRiwrmOutputsModel}(
  x,
  InputsModel,
  RunOptions,
  IndPeriod_Run = which(InputsModel[[1]]$DatesR \%in\% DatesR),
  DatesR = getNextTimeSteps(x),
  Qinf = NULL,
  Qrelease = NULL,
  Qmin = NULL,
  merge_outputs = TRUE,
  ...
)
}
\arguments{
\item{x}{Object returned by \link{RunModel.GRiwrmInputsModel},
\link{RunModel.Supervisor}, or \link{RunModel.GRiwrmOutputsModel}}

\item{InputsModel}{[\code{GRiwrmInputsModel} object] (see
\link{CreateInputsModel.GRiwrm}) or [\code{Supervisor} object] (See \link{CreateSupervisor})}

\item{RunOptions}{[object of class \emph{GRiwrmRunOptions}] see \link{CreateRunOptions.GRiwrmInputsModel} for details}

\item{IndPeriod_Run}{[numeric] index of period to be used for the model run [-]. See details}

\item{DatesR}{(optional) \link{POSIXt} vector of dates of period to be used for
the model run. See details}

\item{Qinf}{(optional) \link{matrix} or \link{data.frame} of \link{numeric} containing
observed flows. It must be provided only for nodes of type "Direct
injection" and "Diversion" [m3 per time step].
Column names correspond to node IDs. Negative flows are abstracted from
the model and positive flows are injected to the model. See details}

\item{Qrelease}{(optional) \link{matrix} or \link{data.frame} of \link{numeric} containing
release flows by nodes using the model \code{RunModel_Reservoir} [m3 per
time step]. See details}

\item{Qmin}{(optional) \link{matrix} or \link{data.frame} of \link{numeric} containing
minimum flows to let downstream of a node with a Diversion [m3 per
time step]. Default is zero. Column names correspond to node IDs.
See details}

\item{merge_outputs}{\link{logical} Merge simulation outputs with the one provided
in argument \code{x}}

\item{...}{Further arguments for compatibility with S3 methods}
}
\value{
An object of class \emph{GRiwrmOutputsModel}.
This object is a \link{list} of \emph{OutputsModel} objects produced by \link{RunModel.InputsModel}
for each node of the semi-distributed model.

It also contains the following attributes (see \link{attr}):
\itemize{
\item "Qm3s": a \link{data.frame} containing the dates of simulation and one column by node
with the simulated flows in cubic meters per seconds (See \link{plot.Qm3s})
\item "GRiwrm":  a copy of the \emph{GRiwrm} object produced by \link{CreateGRiwrm} and used for the simulation
\item "TimeStep":  time step of the simulation in seconds
}
}
\description{
This function allows to restart a simulation at the end of a previous
simulation period. Parameters \code{Qinf}, \code{Qrelease}, and \code{Qmin} can be
redefined for this new simulation period.
}
\details{
\code{IndPeriod_Run} or \code{DatesR} must must be continuous periods starting the
time step after the last simulation time step of the \code{GRiwrmOutputsModel}
object provided through the argument \code{x}.

\code{Qinf}, \code{Qmin}, and \code{Qrelease} are used for overwriting the corresponding
arguments provided to \link{CreateInputsModel.GRiwrm} for the period to be simulated.
Therefore, the number of rows of these arguments must correspond to
\code{IndPeriod_Run} or \code{DatesR} lengths.
}
