% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.GRiwrmOutputsModel.R
\name{RunModel.GRiwrmOutputsModel}
\alias{RunModel.GRiwrmOutputsModel}
\title{RunModel for hot restart after a previous simulation period}
\usage{
\method{RunModel}{GRiwrmOutputsModel}(
  x,
  InputsModel,
  RunOptions,
  IndPeriod_Run = which(InputsModel[[1]]$DatesR \%in\% DatesR),
  DatesR = getNextTimeSteps(x),
  Qinf = NULL,
  Qrelease = NULL,
  Qmin = NULL,
  merge_outputs = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{GRiwrmOutputsModel} object resulting from a previous run.}

\item{InputsModel}{\code{GRiwrmInputsModel} object (see \link{CreateInputsModel.GRiwrm})
or \code{Supervisor} object (see \link{CreateSupervisor}).}

\item{RunOptions}{List of run options created with \link{CreateRunOptions}.}

\item{IndPeriod_Run}{Integer vector indicating the indices of the time steps
to run. Must start at the index immediately following the previous run.}

\item{DatesR}{(optional) \code{POSIXt} vector of dates for the simulation period.
See details.}

\item{Qinf}{(optional) \code{matrix} or \code{data.frame} of \code{numeric} observed flows for
nodes of type "Direct injection" and "Diversion" (m³ per time step).
Column names correspond to node IDs. Negative flows are abstracted from the
model and positive flows are injected to the model. See details.}

\item{Qrelease}{(optional) \code{matrix} or \code{data.frame} of \code{numeric} release flows by
nodes using the model \code{RunModel_Reservoir} (m³ per time step). See details.}

\item{Qmin}{(optional) \code{matrix} or \code{data.frame} of \code{numeric} minimum flows for
downstream of a Diversion node (m³ per time step). Default is zero.
Column names correspond to node IDs. See details.}

\item{merge_outputs}{\code{logical} Merge simulation outputs with the one provided
in argument \code{x}.}

\item{...}{Further arguments for compatibility with S3 methods.}
}
\value{
An object of class \code{GRiwrmOutputsModel} (see \link{RunModel.GRiwrmInputsModel}
for details).
}
\description{
This function restarts a simulation using the state at the end of a previous
simulation (\code{GRiwrmOutputsModel} object \code{x}). It allows redefining the
boundary conditions \code{Qinf}, \code{Qrelease}, and \code{Qmin} for the new period.
}
\details{
\code{IndPeriod_Run} or \code{DatesR} must be continuous periods starting the time step
after the last simulation time step of the \code{GRiwrmOutputsModel} object
provided through the argument \code{x}.

\code{Qinf}, \code{Qmin}, and \code{Qrelease} are used to overwrite the corresponding
arguments provided to \link{CreateInputsModel.GRiwrm} for the period to be
simulated. Therefore, the number of rows of these arguments must correspond
to \code{IndPeriod_Run} or \code{DatesR} lengths.
}
\seealso{
\code{\link[=CreateGRiwrm]{CreateGRiwrm()}}, \code{\link[=CreateInputsModel.GRiwrm]{CreateInputsModel.GRiwrm()}}, \code{\link[=CreateRunOptions]{CreateRunOptions()}}

Vignette "V07_Combine_tactical_operational_management" in package airGRiwrm
}
