% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModel.Supervisor.R
\name{RunModel.Supervisor}
\alias{RunModel.Supervisor}
\title{RunModel function for a GRiwrmInputsModel object}
\usage{
\method{RunModel}{Supervisor}(x, RunOptions, Param, ...)
}
\arguments{
\item{x}{[object of class \code{Supervisor}] see \link{CreateSupervisor} for details}

\item{RunOptions}{[object of class \emph{GRiwrmRunOptions}] see \code{[CreateRunOptions.GRiwrm]} for details}

\item{Param}{\link{list} parameter values. The list item names are the IDs of the sub-basins. Each item is a vector of numerical parameters}

\item{...}{Further arguments for compatibility with S3 methods}
}
\value{
\emph{GRiwrmOutputsModel} object which is a list of \emph{OutputsModel} objects (See \link[airGR:RunModel]{airGR::RunModel}) for each node of the semi-distributed model
}
\description{
RunModel function for a GRiwrmInputsModel object
}
