% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setDataToLocation}
\alias{setDataToLocation}
\title{Writing of data to model input for the current time step}
\usage{
setDataToLocation(ctrlr, sv)
}
\arguments{
\item{ctrlr}{[object of type \code{Controller}] see \link{CreateController} for details}

\item{sv}{[object of type \code{Supervisor}] see \link{CreateSupervisor} for details}
}
\value{
\link{NULL}
}
\description{
Writing of data to model input for the current time step
}
