% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.Calibration.R
\name{transferGRparams}
\alias{transferGRparams}
\title{Transfer GR parameters from one donor sub-basin to a receiver sub-basin}
\usage{
transferGRparams(
  InputsModel,
  Param,
  donor,
  receiver,
  default_param = NULL,
  verbose = FALSE
)
}
\arguments{
\item{InputsModel}{Object of class \link[=CreateInputsModel.GRiwrm]{GRiwrmInputsModel},
see \link{CreateInputsModel.GRiwrm} for details}

\item{Param}{\link{numeric} vector of GR model parameters}

\item{donor}{\link{character} id of the node which gives its parameters}

\item{receiver}{\link{character} id of the node which receives the parameters from the donor}

\item{default_param}{\link{numeric} vector of GR model parameters if parameters are missing from the donor}

\item{verbose}{\link{logical} Add information message on donor and receiver}
}
\value{
A \link{numeric} \link{vector} of transferred parameters
}
\description{
This function is used by \code{Calibration.GRiwrmInputsModel} for transferring parameters
to ungauged nodes and for providing a priori parameters in parameter
regularization (See \link{CreateInputsCrit}).
}
\details{
\code{donor} and \code{receiver} nodes should have the same GR model with the same snow
module configuration.

The transfer takes care of:
\itemize{
\item the presence/absence of hydraulic routing parameters between the donor and
the receiver (But \code{default_param} should be provided if the receiver has more
parameters than the donor)
\item the transformation of the X4 parameters of GR models. Indeed, this parameter
is correlated to the catchment area.
Therefore, it has to be rescaled from the surface of the donor catchment to
the surface of the receiver catchment, following the relationship suggested
by Lobligeois (2014):
\eqn{X4_{receiver} = X4_{donor} \times (S_{receiver} / S_{donor})^{0.3}}.
}
}
\references{
Lobligeois, F., 2014. Mieux connaitre la distribution spatiale des pluies améliore-t-il la modélisation des crues ?
Diagnostic sur 181 bassins versants français. Thèse de Doctorat, Irstea (Antony), AgroParisTech (Paris), 312 pp.
https://webgr.inrae.fr/Media/Files/biblio-theses/2014_lobligeois_these.pdf
}
