% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateQsimUpstream.R
\name{UpdateQsimUpstream}
\alias{UpdateQsimUpstream}
\title{Update InputsModel$Qupstream with simulated upstream flows provided by GRiwrmOutputsModels object.}
\usage{
UpdateQsimUpstream(InputsModel, IndPeriod_Run, OutputsModel)
}
\arguments{
\item{InputsModel}{\emph{InputsModel} object. See \code{\link[airGR]{CreateInputsModel}}}

\item{IndPeriod_Run}{numeric index of period to be used for the model run (-)}

\item{OutputsModel}{\emph{GRiwrmOutputsModel} object provided by \link{RunModel.GRiwrmInputsModel}.}
}
\value{
\code{InputsModel} object with updated \code{Qupstream}
}
\description{
This function is used by \link{RunModel.GRiwrmInputsModel} and \link{Calibration.GRiwrmInputsModel} in order to provide upstream simulated flows to a node.
}
